@@ -1,12 +1,12 @@
 module github.com/uselagoon/lagoon/services/backup-handler
 
-go 1.18
+go 1.21
 
 require (
-	github.com/amazeeio/lagoon-cli v0.0.0-20200309230012-d971a4d07053
 	github.com/google/uuid v1.3.0
 	github.com/isayme/go-amqp-reconnect v0.0.0-20210303120416-fc811b0bcda2
 	github.com/streadway/amqp v1.0.0
+	github.com/uselagoon/lagoon-cli v0.19.0
 )
 
 require (
