@@ -120,28 +120,6 @@ const apolloServer = new ApolloServer({
         esClient,
       };
 
-      // get all keycloak groups, do this early to reduce the number of times this is called otherwise
-      // but doing this early and once is pretty cheap
-      let keycloakGroups = []
-      try {
-        // check redis for the allgroups cache value
-        const data = await getRedisKeycloakCache("allgroups");
-        let buff = new Buffer(data, 'base64');
-        keycloakGroups = JSON.parse(buff.toString('utf-8'));
-      } catch (err) {
-        logger.warn(`Couldn't check redis keycloak cache: ${err.message}`);
-        // if it can't be recalled from redis, get the data from keycloak
-        const allGroups = await Group.Group(modelClients).loadAllGroups();
-        keycloakGroups = await Group.Group(modelClients).transformKeycloakGroups(allGroups);
-        const data = Buffer.from(JSON.stringify(keycloakGroups)).toString('base64')
-        try {
-          // then attempt to save it to redis
-          await saveRedisKeycloakCache("allgroups", data);
-        } catch (err) {
-          logger.warn(`Couldn't save redis keycloak cache: ${err.message}`);
-        }
-      }
-
       let currentUser = {};
       let serviceAccount = {};
       // if this is a user request, get the users keycloak groups too, do this one to reduce the number of times it is called elsewhere
@@ -149,12 +127,14 @@ const apolloServer = new ApolloServer({
       let keycloakUsersGroups = []
       let groupRoleProjectIds = []
       const keycloakGrant = grant
+      let legacyGrant = legacyCredentials ? legacyCredentials : null
       if (keycloakGrant) {
+        // get all the users keycloak groups, do this early to reduce the number of times this is called otherwise
         keycloakUsersGroups = await User.User(modelClients).getAllGroupsForUser(keycloakGrant.access_token.content.sub);
         serviceAccount = await keycloakGrantManager.obtainFromClientCredentials();
         currentUser = await User.User(modelClients).loadUserById(keycloakGrant.access_token.content.sub);
         // grab the users project ids and roles in the first request
-        groupRoleProjectIds = await User.User(modelClients).getAllProjectsIdsForUser(currentUser, keycloakUsersGroups);
+        groupRoleProjectIds = await User.User(modelClients).getAllProjectsIdsForUser(currentUser.id, keycloakUsersGroups);
       }
 
       return {
@@ -164,6 +144,7 @@ const apolloServer = new ApolloServer({
           ? keycloakHasPermission(grant, requestCache, modelClients, serviceAccount, currentUser, groupRoleProjectIds)
           : legacyHasPermission(legacyCredentials),
         keycloakGrant,
+        legacyGrant,
         requestCache,
         models: {
           UserModel: User.User(modelClients),
@@ -171,7 +152,6 @@ const apolloServer = new ApolloServer({
           ProjectModel: ProjectModel.ProjectModel(modelClients),
           EnvironmentModel: EnvironmentModel.EnvironmentModel(modelClients)
         },
-        keycloakGroups,
         keycloakUsersGroups,
       };
     },
@@ -204,28 +184,6 @@ const apolloServer = new ApolloServer({
         esClient,
       };
 
-      // get all keycloak groups, do this early to reduce the number of times this is called otherwise
-      // but doing this early and once is pretty cheap
-      let keycloakGroups = []
-      try {
-        // check redis for the allgroups cache value
-        const data = await getRedisKeycloakCache("allgroups");
-        let buff = new Buffer(data, 'base64');
-        keycloakGroups = JSON.parse(buff.toString('utf-8'));
-      } catch (err) {
-        logger.warn(`Couldn't check redis keycloak cache: ${err.message}`);
-        // if it can't be recalled from redis, get the data from keycloak
-        const allGroups = await Group.Group(modelClients).loadAllGroups();
-        keycloakGroups = await Group.Group(modelClients).transformKeycloakGroups(allGroups);
-        const data = Buffer.from(JSON.stringify(keycloakGroups)).toString('base64')
-        try {
-          // then attempt to save it to redis
-          await saveRedisKeycloakCache("allgroups", data);
-        } catch (err) {
-          logger.warn(`Couldn't save redis keycloak cache: ${err.message}`);
-        }
-      }
-
       let currentUser = {};
       let serviceAccount = {};
       // if this is a user request, get the users keycloak groups too, do this one to reduce the number of times it is called elsewhere
@@ -233,12 +191,14 @@ const apolloServer = new ApolloServer({
       let keycloakUsersGroups = []
       let groupRoleProjectIds = []
       const keycloakGrant = req.kauth ? req.kauth.grant : null
+      let legacyGrant = req.legacyCredentials ? req.legacyCredentials : null
       if (keycloakGrant) {
+        // get all the users keycloak groups, do this early to reduce the number of times this is called otherwise
         keycloakUsersGroups = await User.User(modelClients).getAllGroupsForUser(keycloakGrant.access_token.content.sub);
         serviceAccount = await keycloakGrantManager.obtainFromClientCredentials();
         currentUser = await User.User(modelClients).loadUserById(keycloakGrant.access_token.content.sub);
         // grab the users project ids and roles in the first request
-        groupRoleProjectIds = await User.User(modelClients).getAllProjectsIdsForUser(currentUser, keycloakUsersGroups);
+        groupRoleProjectIds = await User.User(modelClients).getAllProjectsIdsForUser(currentUser.id, keycloakUsersGroups);
       }
 
       // do a permission check to see if the user is platform admin/owner, or has permission for `viewAll` on certain resources
@@ -284,6 +244,7 @@ const apolloServer = new ApolloServer({
         hasPermission,
         keycloakGrant,
         requestCache,
+        legacyGrant,
         userActivityLogger: (message, meta) => {
           let defaultMeta = {
             user: req.kauth
@@ -304,7 +265,6 @@ const apolloServer = new ApolloServer({
           ProjectModel: ProjectModel.ProjectModel(modelClients),
           EnvironmentModel: EnvironmentModel.EnvironmentModel(modelClients)
         },
-        keycloakGroups,
         keycloakUsersGroups,
         adminScopes: {
           projectViewAll: projectViewAll,
@@ -374,9 +334,9 @@ const apolloServer = new ApolloServer({
 
         // operationName is set by the client and optional. rootFieldName is
         // set by the API type defs.
-        // operationName would be "getHighCottonProjectId" and rootFieldName
+        // operationName would be "getLagoonDemoProjectId" and rootFieldName
         // would be "getProjectByName" with a query like:
-        // query getHighCottonProjectId { getProjectByName(name: "high-cotton") { id } }
+        // query getLagoonDemoProjectId { getProjectByName(name: "lagoon-demo") { id } }
         const transactionName = operationName ? operationName : rootFieldName;
         newrelic.setTransactionName(`graphql (${transactionName})`);
         newrelic.addCustomAttribute('gqlQuery', queryString);
