@@ -7,9 +7,6 @@ import { ResolverFn } from '../index';
 import { knex } from '../../util/db';
 import { logger } from '../../loggers/logger';
 import crypto from 'crypto';
-import { Service } from 'aws-sdk';
-import * as api from '@lagoon/commons/dist/api';
-import { getEnvironmentsByProjectId } from '../environment/resolvers';
 import { getUserProjectIdsFromRoleProjectIds } from '../../util/auth';
 
 export const getFactsByEnvironmentId: ResolverFn = async (
@@ -131,9 +128,7 @@ export const getProjectsByFactSearch: ResolverFn = async (
   let userProjectIds: number[];
 
   if (!adminScopes.projectViewAll) {
-    const userProjectRoles = await models.UserModel.getAllProjectsIdsForUser({
-      id: keycloakGrant.access_token.content.sub
-    }, keycloakUsersGroups);
+    const userProjectRoles = await models.UserModel.getAllProjectsIdsForUser(keycloakGrant.access_token.content.sub, keycloakUsersGroups);
     userProjectIds = getUserProjectIdsFromRoleProjectIds(userProjectRoles);
   }
 
@@ -151,9 +146,7 @@ export const getEnvironmentsByFactSearch: ResolverFn = async (
 
   let userProjectIds: number[];
   if (!adminScopes.projectViewAll) {
-    const userProjectRoles = await models.UserModel.getAllProjectsIdsForUser({
-      id: keycloakGrant.access_token.content.sub
-    }, keycloakUsersGroups);
+    const userProjectRoles = await models.UserModel.getAllProjectsIdsForUser(keycloakGrant.access_token.content.sub, keycloakUsersGroups);
     userProjectIds = getUserProjectIdsFromRoleProjectIds(userProjectRoles);
   }
 
@@ -331,7 +324,7 @@ export const addFactsByName: ResolverFn = async (
   }
 
   const returnFacts = await processAddFacts(
-    R.map((fact) => {return {environment: envId, ...fact};}, facts),
+    R.map((fact:object) => {return {environment: envId, ...fact};}, facts),
     sqlClientPool,
     hasPermission,
     adminScopes,
@@ -382,18 +375,22 @@ export const deleteFact: ResolverFn = async (
 
 export const deleteFactsFromSource: ResolverFn = async (
   root,
-  { input: { environment: environmentId, source } },
+  { input: { environment: environmentId, source, service } },
   { sqlClientPool, hasPermission, userActivityLogger }
 ) => {
   const environment = await environmentHelpers(
     sqlClientPool
   ).getEnvironmentById(environmentId);
 
+  if(environment == null) {
+    throw new Error(`Unable to find environment with id: ${environmentId}`);
+  }
+
   await hasPermission('fact', 'delete', {
     project: environment.project
   });
 
-  await query(sqlClientPool, Sql.deleteFactsFromSource(environmentId, source));
+  await query(sqlClientPool, Sql.deleteFactsFromSource(environmentId, source, service));
 
   userActivityLogger(`User deleted facts`, {
     project: '',
@@ -401,7 +398,8 @@ export const deleteFactsFromSource: ResolverFn = async (
     payload: {
       data: {
         environment: environmentId,
-        source
+        source,
+        service
       }
     }
   });
