@@ -48,11 +48,25 @@ export const Sql = {
       })
       .del()
       .toString(),
-  deleteFactsFromSource: (environment, source) =>
-    knex('environment_fact')
-      .where({ environment, source })
-      .del()
-      .toString(),
+  deleteFactsForEnvironment: (environment) => // Used when removing environments.
+      knex('environment_fact')
+        .where({
+          environment
+        })
+        .del()
+        .toString(),
+  deleteFactsFromSource: (environment, source, service) => {
+    let query =  knex('environment_fact')
+      .where({ environment, source });
+
+      if(service != null) {
+        query = query.andWhere("service", "=", service)
+      }
+
+    query = query.del();
+
+    return query.toString();
+  },
   selectFactReferenceByDatabaseId: (id: number) =>
     knex('environment_fact_reference')
       .where({ id })
