@@ -215,8 +215,8 @@ export const addProblem: ResolverFn = async (
 
 export const deleteProblem: ResolverFn = async (
   root,
-  { input: { environment: environmentId, identifier } },
-  { sqlClientPool, hasPermission, userActivityLogger }
+  { input: { environment: environmentId, identifier, service } },
+  { sqlClientPool, hasPermission, userActivityLogger  }
 ) => {
   const environment = await environmentHelpers(
     sqlClientPool
@@ -226,7 +226,7 @@ export const deleteProblem: ResolverFn = async (
     project: environment.project
   });
 
-  await query(sqlClientPool, Sql.deleteProblem(environmentId, identifier));
+  await query(sqlClientPool, Sql.deleteProblem(environmentId, identifier, service));
 
   userActivityLogger(`User deleted a problem on environment '${environment.name}' for '${environment.project}'`, {
     project: '',
@@ -273,14 +273,7 @@ export const getProblemHarborScanMatches: ResolverFn = async (
   args,
   { sqlClientPool, hasPermission }
 ) => {
-  await hasPermission('harbor_scan_match', 'view', {});
-
-  const rows = await query(
-    sqlClientPool,
-    Sql.selectAllProblemHarborScanMatches()
-  );
-
-  return rows;
+  throw new Error('Harbor-Trivy integration with core removed in Lagoon 2')
 };
 
 export const addProblemHarborScanMatch: ResolverFn = async (
@@ -297,42 +290,7 @@ export const addProblemHarborScanMatch: ResolverFn = async (
   },
   { sqlClientPool, hasPermission, userActivityLogger }
 ) => {
-  await hasPermission('harbor_scan_match', 'add', {});
-
-  const { insertId } = await query(
-    sqlClientPool,
-    Sql.insertProblemHarborScanMatch({
-      id: null,
-      name,
-      description,
-      default_lagoon_project: defaultLagoonProject,
-      default_lagoon_environment: defaultLagoonEnvironment,
-      default_lagoon_service_name: defaultLagoonService,
-      regex
-    })
-  );
-
-  const rows = await query(
-    sqlClientPool,
-    Sql.selectAllProblemHarborScanMatchByDatabaseId(insertId)
-  );
-
-  userActivityLogger(`User added harbor scan regex matcher`, {
-    project: '',
-    event: 'api:addProblemHarborScanMatch',
-    payload: {
-      input: {
-        name,
-        description,
-        defaultLagoonProject,
-        defaultLagoonEnvironment,
-        defaultLagoonService,
-        regex
-      }
-    }
-  });
-
-  return R.prop(0, rows);
+  throw new Error('Harbor-Trivy integration with core removed in Lagoon 2')
 };
 
 export const deleteProblemHarborScanMatch: ResolverFn = async (
@@ -340,17 +298,5 @@ export const deleteProblemHarborScanMatch: ResolverFn = async (
   { input: { id } },
   { sqlClientPool, hasPermission, userActivityLogger }
 ) => {
-  await hasPermission('harbor_scan_match', 'delete', {});
-
-  await query(sqlClientPool, Sql.deleteProblemHarborScanMatch(id));
-
-  userActivityLogger(`User deleted harbor scan regex matcher`, {
-    project: '',
-    event: 'api:deleteProblemHarborScanMatch',
-    payload: {
-      input: { id }
-    }
-  });
-
-  return 'success';
+  throw new Error('Harbor-Trivy integration with core removed in Lagoon 2')
 };
