@@ -49,7 +49,7 @@ export const addSshKey: ResolverFn = async (
 
   const user = await models.UserModel.loadUserByIdOrUsername({
     id: R.prop('id', userInput),
-    username: R.prop('email', userInput)
+    email: R.prop('email', userInput)
   });
 
   await hasPermission('ssh_key', 'add', {
