@@ -1,18 +1,16 @@
-// @ts-ignore
 import * as R from 'ramda';
 import { verify } from 'jsonwebtoken';
 import { logger } from '../loggers/logger';
 import { getConfigFromEnv } from '../util/config';
 import { isNotNil } from './func';
 import { keycloakGrantManager } from '../clients/keycloakClient';
-// @ts-ignore
 const { userActivityLogger } = require('../loggers/userActivityLogger');
 import { Group } from '../models/group';
 import { User } from '../models/user';
-import { saveRedisCache, getRedisCache, saveRedisKeycloakCache } from '../clients/redisClient';
 
 interface ILegacyToken {
   iat: string;
+  exp: string;
   iss: string;
   sub: string;
   aud: string;
@@ -105,7 +103,26 @@ export const getCredentialsForLegacyToken = async token => {
     throw new Error('Decoding token resulted in "null" or "undefined".');
   }
 
-  const { role = 'none', aud, sub, iss, iat } = decoded;
+  const { role = 'none', aud, sub, iss, iat, exp } = decoded;
+
+  // check the expiration on legacy tokens, reject them if necessary
+  const maxExpiry = getConfigFromEnv('LEGACY_EXPIRY_MAX', '3600') // 1hour default
+  const rejectLegacyExpiry = getConfigFromEnv('LEGACY_EXPIRY_REJECT', 'false') // don't reject intially, just log
+  if (exp) {
+    if ((parseInt(exp)-parseInt(iat)) > parseInt(maxExpiry)) {
+      const msg = `Legacy token (sub:${sub}; iss:${iss}) expiry ${(parseInt(exp)-parseInt(iat))} is greater than ${parseInt(maxExpiry)}`
+      logger.warn(msg);
+      if (rejectLegacyExpiry == "true") {
+        throw new Error(msg);
+      }
+    }
+  } else {
+    const msg = `Legacy token (sub:${sub}; iss:${iss}) has no expiry`
+    logger.warn(msg);
+    if (rejectLegacyExpiry == "true") {
+      throw new Error(msg);
+    }
+  }
 
   if (aud !== getConfigFromEnv('JWTAUDIENCE')) {
     throw new Error('Token audience mismatch.');
@@ -154,54 +171,6 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
   const UserModel = User(modelClients);
 
   return async (resource, scope, attributes: IKeycloakAuthAttributes = {}) => {
-
-    // Check if the same set of permissions has been granted already for this
-    // api query.
-    const cacheKey = `${currentUser.id}:${resource}:${scope}:${JSON.stringify(
-      attributes
-    )}`;
-    const cachedPermissions = requestCache.get(cacheKey);
-    if (cachedPermissions === true) {
-      return true;
-    } else if (!cachedPermissions === false) {
-      userActivityLogger.user_info(
-        `User does not have permission to '${scope}' on '${resource}'`,
-        {
-          user: grant ? grant.access_token.content : null
-        }
-      );
-      throw new KeycloakUnauthorizedError(
-        `Unauthorized: You don't have permission to "${scope}" on "${resource}": ${JSON.stringify(
-          attributes
-        )}`
-      );
-    }
-
-    // Check the redis cache before doing a full keycloak lookup.
-    const resourceScope = { resource, scope, currentUserId: currentUser.id, ...attributes };
-    let redisCacheResult: number;
-    try {
-      const data = await getRedisCache(resourceScope);
-      redisCacheResult = parseInt(data, 10);
-    } catch (err) {
-      logger.warn(`Couldn't check redis authz cache: ${err.message}`);
-    }
-
-    if (redisCacheResult === 1) {
-      return true;
-    } else if (redisCacheResult === 0) {
-      userActivityLogger.user_info(
-        `User does not have permission to '${scope}' on '${resource}'`,
-        {
-          user: grant.access_token.content
-        }
-      );
-      throw new KeycloakUnauthorizedError(
-        `Unauthorized: You don't have permission to "${scope}" on "${resource}": ${JSON.stringify(
-          attributes
-        )}`
-      );
-    }
 
     let claims: {
       currentUser: [string];
@@ -272,7 +241,7 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
           // but could happen elsewhere
           const keycloakUsersGroups = await UserModel.getAllGroupsForUser(currentUser.id);
           // grab the users project ids and roles in the first request
-          groupRoleProjectIds = await UserModel.getAllProjectsIdsForUser(currentUser, keycloakUsersGroups);
+          groupRoleProjectIds = await UserModel.getAllProjectsIdsForUser(currentUser.id, keycloakUsersGroups);
 
           [highestRoleForProject, upids] = getUserRoleForProjectFromRoleProjectIds(groupRoleProjectIds, projectId)
         }
@@ -354,7 +323,6 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
       authzRequest = {
         ...authzRequest,
         claim_token_format: 'urn:ietf:params:oauth:token-type:jwt',
-        // @ts-ignore
         claim_token: Buffer.from(JSON.stringify(claims)).toString('base64')
       };
     }
@@ -367,19 +335,12 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
     };
 
     try {
-      // @ts-ignore
       const newGrant = await keycloakGrantManager.checkPermissions(
         authzRequest,
         request
       );
 
       if (newGrant.access_token.hasPermission(resource, scope)) {
-        requestCache.set(cacheKey, true);
-        try {
-          await saveRedisCache(resourceScope, '1');
-        } catch (err) {
-          logger.warn(`Couldn't save redis authz cache: ${err.message}`);
-        }
         return;
       }
     } catch (err) {
