@@ -1,10 +1,5 @@
 // @flow
 
-import {
-  addProblem,
-  deleteProblemsFromSource,
-  getProblemsforProjectEnvironment,
-} from'@lagoon/commons/dist/api';
 import { sendToLagoonLogs } from '@lagoon/commons/dist/logs/lagoon-logger';
 const DRUTINY_VULNERABILITY_SOURCE_BASE = 'Drutiny';
 const DRUTINY_SERVICE_NAME = 'cli';
@@ -13,6 +8,9 @@ import {
   getProjectByName,
   getEnvironmentByOpenshiftProjectName,
   getOpenShiftInfoForProject,
+  addProblem,
+  deleteProblemsFromSource,
+  getProblemsforProjectEnvironment,
 } from '@lagoon/commons/dist/api';
 import { generateProblemsWebhookEventName } from "./webhookHelpers";
 import * as R from 'ramda';
