@@ -1,53 +1,55 @@
+import { Project as LagoonProject } from '@lagoon/commons/dist/api';
+
 /**
  * Webhook related types for incoming Events
  */
 
 export interface removeData {
-  projectName: string,
-  pullrequestNumber?: number,
-  pullrequestTitle?: string,
-  openshiftProjectName?: string,
-  forceDeleteProductionEnvironment?: boolean,
-  branchName?: string,
-  branch?: string,
-  pullrequest?: string,
-  type: string
-};
+  branch?: string;
+  branchName?: string;
+  forceDeleteProductionEnvironment?: boolean;
+  openshiftProjectName?: string;
+  projectName: string;
+  pullrequest?: string;
+  pullrequestNumber?: number;
+  pullrequestTitle?: string;
+  type: string;
+}
 
 export interface deployData {
-  projectName: string,
-  branchName: string,
-  sha?: string,
-  repoUrl?: string,
-  repoName?: string,
-  pullrequestTitle?: string,
-  pullrequestNumber?: number,
-  pullrequestUrl?: string,
-  type?: string,
-  headBranchName?: string,
-  headSha?: string,
-  baseBranchName?: string,
-  baseSha?: string,
-  buildName?: string,
-  buildPriority?: string,
-  bulkId?: string,
-  buildVariables?: any,
-};
+  baseBranchName?: string;
+  baseSha?: string;
+  branchName: string;
+  buildName?: string;
+  buildPriority?: string;
+  buildVariables?: any;
+  bulkId?: string;
+  headBranchName?: string;
+  headSha?: string;
+  projectName: string;
+  pullrequestNumber?: number;
+  pullrequestTitle?: string;
+  pullrequestUrl?: string;
+  repoName?: string;
+  repoUrl?: string;
+  sha?: string;
+  sourceType?: string,
+  sourceUser?: string,
+  type?: string;
+}
 
 export interface WebhookRequestData {
-  webhooktype: string,
-  event: string,
-  giturl: string,
-  uuid?: string,
-  body?: any,
-  user?: any,
-  sender?: any
-};
+  body?: any;
+  event: string;
+  giturl: string;
+  sender?: any;
+  user?: any;
+  uuid?: string;
+  webhooktype: string;
+}
 
-export interface Project {
-  slack: any,
-  name: string,
-  deploymentsDisabled: number,
-  openshift: any,
-  productionEnvironment?: string
-};
+export type Project = Pick<
+  LagoonProject,
+  | 'name'
+  | 'deploymentsDisabled'
+>;
